Cereal

Chakrapani and Ehrenberg [1981] analyzed people's attitudes towards a variety of breakfast cereals. The data matrix cereal is 8 ? 11, with rows corresponding to eight cereals, and columns corresponding to potential attributes about cereals. The original data consisted of the percentage of subjects who thought the given cereal possessed the given attribute. The present matrix has been doubly centered, so that the row means and columns means are all zero. (The original data can be found in the S-Plus [TIBCO Software Inc., 2009] data set cereal.attitude.)

cereal

Format

A double matrix with 8 observations on the following 11 variables.

Return

A cereal one would come back to

Tasty

Tastes good

Popular

Popular with the entire family

Nourishing

Cereal is fulfilling

NaturalFlavor

Cereal lacks flavor additives

Affordable

Cereal is priced well for the content

GoodValue

Quantity for Price

Crispy

Stays crispy in milk

Fit

Keeps one fit

Fun

Fun for children

Source

T. K. Chakrapani and A. S. C. Ehrenberg. An alternative to factor analysis in marketing research part 2: Between group analysis. Professional Marketing Research Society Journal, 1:32-38, 1981.