Morphological Measurements on Leptograpsus Crabs

The crabs data frame has 200 rows and 8 columns, describing 5 morphological measurements on 50 crabs each of two colour forms and both sexes, of the species Leptograpsus variegatus collected at Fremantle, W. Australia.

crabs

Format

This data frame contains the following columns:

sp

species - "B" or "O" for blue or orange.

sex

"M" (Male) or "F" (Female).

index

index 1:50 within each of the four groups.

FL

frontal lobe size (mm).

RW

rear width (mm).

CL

carapace length (mm).

CW

carapace width (mm).

BD

body depth (mm).

Source

Campbell, N.A. and Mahon, R.J. (1974) A multivariate study of variation in two species of rock crab of genus Leptograpsus. Australian Journal of Zoology 22, 417--425.

MASS, R-Package

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.