Maximizing negentropy for \(q = 2\) dimensions

Searches for the rotation that maximizes the estimated negentropy of the first column of the rotated data, for \(q = 2\) dimensional data.

negent2D(y, m = 100)

Arguments

y
The \(n \times 2\) data matrix.

m
The number of angles (between \(0\) and \(\pi\)) over which to search.

Value

A list with the following components:

vectors

The \(2 ? 2\) orthogonal matrix G that optimizes the negentropy.

values

Estimated negentropies for the two rotated variables. The largest is first.

See also

Examples

# Load iris data
data(iris)

# Centers and scales the variables.
y <- scale(as.matrix(iris[, 1:2]))

# Obtains Negent Vectors for 2 x 2 matrix
gstar <- negent2D(y, m = 10)$vectors