BIC for a particular pattern

Find the BIC and MLE from a set of observed eigenvalues for a specific pattern.

pcbic(eigenvals, n, pattern)

Arguments

eigenvals

The \(Q\)-vector of eigenvalues of the covariance matrix, in order from largest to smallest.

n

The degrees of freedom in the covariance matrix.

pattern

The pattern of equalities of the eigenvalues, given by the \(K\)-vector (\(Q_1\), ... , \(Q_K\)) as in (13.8).

Value

A `list` with the following components:

lambdaHat

A \(Q\)-vector containing the MLE's for the eigenvalues.

Deviance

The deviance of the model, as in (13.13).

Dimension

The dimension of the model, as in (13.12).

BIC

The value of the BIC for the model, as in (13.14).

See also

Examples

# Build cars1
require("mclust")
mcars <- Mclust(cars)
cars1 <- cars[mcars$classification == 1, ]
xcars <- scale(cars1)
eg <- eigen(var(xcars))
pcbic(eg$values, 95, c(1, 1, 3, 3, 2, 1))
#> $lambdaHat
#>  [1] 6.20990339 1.83332778 0.71561848 0.71561848 0.71561848 0.21276067
#>  [7] 0.21276067 0.21276067 0.07077652 0.07077652 0.03007833
#> 
#> $Deviance
#> [1] -1141.398
#> 
#> $Dimension
#> [1] 54
#> 
#> $BIC
#> [1] -895.4885
#>